third_party/WebKit/LayoutTests/media/video-force-preload-none-to-metadata-on-load.html

<!DOCTYPE html>
<title>Test that loading a video element with preload "none" changes its effective preload to "metadata"</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video preload="none"></video>
<script>
test(function() {
    var video = document.querySelector('video');
    video.src = 'content/test.ogv';
    assert_equals(internals.effectivePreload(video), "none");
    video.load();
    assert_equals(internals.effectivePreload(video), "metadata");
});
</script>
